backend/requirements.txt
flask
transformers
torch
newspaper3k
flask_cors